fix: fix two
