+++ web/src/Setting.js
@@ -1554,3 +1554,14 @@ export function getCurrencyText(product) {
     return "(Unknown currency)";
   }
 }
+
+const characters = "123456789abcdef";
+
+export function getRandomHexKey(length) {
+  let key = "";
+  for (let i = 0; i < length; i++) {
+    const randomIndex = Math.floor(Math.random() * characters.length);
+    key += characters[randomIndex];
+  }
+  return key;
+}
